+++ hotel-booking-app-service/src/test/java/com/application/hotelbooking/services/repositoryservices/ConfirmationTokenRepositoryServiceImplTest.java
@@ -1,4 +1,17 @@
-package com.application.hotelbooking.services;
+package com.application.hotelbooking.services.repositoryservices;
 
-public class ConfirmationTokenServiceImplTest {
+import com.application.hotelbooking.repositories.ConfirmationTokenRepository;
+import com.application.hotelbooking.transformers.ConfirmationTokenTransformer;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+
+@ExtendWith(MockitoExtension.class)
+public class ConfirmationTokenRepositoryServiceImplTest {
+
+    @Autowired
+    private ConfirmationTokenRepository confirmationTokenRepository;
+
+    @Autowired
+    private ConfirmationTokenTransformer confirmationTokenTransformer;
 }
